Add support for numpy 2.x

+++ audiomentations/augmentations/limiter.py
@@ -108,7 +108,7 @@ class Limiter(BaseWaveformTransform):
 
             self.parameters[
                 "threshold"
-            ] = threshold_factor * convert_decibels_to_amplitude_ratio(threshold_db)
+            ] = float(threshold_factor * convert_decibels_to_amplitude_ratio(threshold_db))
 
     def apply(self, samples: NDArray[np.float32], sample_rate: int) -> NDArray[np.float32]:
         if self.parameters["threshold"] == 0.0:
